Update Bus Automation.py
This is an automatic bus scheduling project for which we have a dataset having different busses numbers and their start time respectively. As the time arrives for any of the busses mentioned there, to deport it will send a notification to your device mentioning the bus number. After that it automatically schedules the time for the next bus to deport